Renamed CommandLineRunner classes, added RoleCommandLineRunner, modified Role entity, added getName for Role service and repository.

+++ src/main/java/com/tmjonker/burgerbonanza/entities/role/Role.java
@@ -8,7 +8,7 @@ public class Role {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(nullable = false)
     private String name;
@@ -19,7 +19,7 @@ public class Role {
         this.name = name;
     }
 
-    public Role(int id, String name ) {
+    public Role(Long id, String name ) {
         this.id = id;
         this.name = name;
     }
@@ -32,11 +32,11 @@ public class Role {
         this.name = name;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
